refactor: extract and consolidate reader UI components into `ReaderScaffold` for improved reusability and maintainability
